backend/requirements.txt
google-adk>=1.27.1
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.25.0
